[PM-19951] Move EncryptedPreferences to the correct di package
The `EncryptedPreferences` annotation has been moved to package `com.bitwarden.data.datasource.disk.di`.
