--- data/src/main/kotlin/com/bitwarden/data/datasource/di/EncryptedPreferences.kt
@@ -1,11 +1,0 @@
-package com.bitwarden.data.datasource.di
-
-import android.content.SharedPreferences
-import javax.inject.Qualifier
-
-/**
- * Used to denote an instance of [SharedPreferences] that encrypts its data.
- */
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class EncryptedPreferences
